move id mapper to its dedicated module
